[fix] production errors
