+++ app/governance/(components)/myproposal/MyProposals.module.scss
@@ -1,0 +1,198 @@
+// SCSS styles for MyProposals component
+// These styles are extracted from the main SCSS file
+
+.myProposalsSection {
+  margin-top: 1rem;
+  width: 100%;
+  background-color: #f5f7ff;
+  border-radius: 2rem;
+  padding: 2rem;
+}
+
+.sectionTitle {
+  font-size: 1.75rem;
+  font-weight: 700;
+  margin-bottom: 1rem;
+  width: fit-content !important;
+}
+
+.card {
+  background: #fff;
+  padding: 1.5rem;
+  border-radius: 1rem;
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.05);
+  border-left: 4px solid #0070f3;
+}
+
+.cardHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  margin-bottom: 16px;
+
+  .proposalTitle {
+    margin: 0;
+    font-size: 20px;
+    font-weight: 600;
+  }
+
+  .proposalAuthor {
+    font-size: 12px;
+    color: #6b7280;
+    margin-top: 4px;
+  }
+
+  .badge {
+    background-color: #ffe69a;
+    color: #a97100;
+    font-size: 12px;
+    font-weight: 500;
+    padding: 4px 10px;
+    border-radius: 6px;
+    height: fit-content;
+  }
+}
+
+.cardMeta {
+  display: flex;
+  gap: 40px;
+  margin-bottom: 20px;
+
+  .metaItem {
+    display: flex;
+    flex-direction: column;
+
+    .metaLabel {
+      font-size: 14px;
+      color: #6b7280;
+    }
+
+    .metaValue {
+      font-size: 16px;
+      font-weight: 500;
+    }
+  }
+}
+
+.progressSection {
+  margin-bottom: 20px;
+
+  .progressLabel {
+    font-size: 12px;
+    color: #6b7280;
+    margin-bottom: 6px;
+  }
+
+  .progressBar {
+    height: 4px;
+    background-color: #e5e7eb;
+    border-radius: 4px;
+    overflow: hidden;
+
+    .filled {
+      width: 50%;
+      height: 100%;
+      background-color: #3b82f6;
+      transition: width 0.3s ease;
+    }
+  }
+
+  .progressSteps {
+    display: flex;
+    justify-content: space-between;
+    font-size: 12px;
+    margin-top: 4px;
+
+    .currentStep {
+      font-weight: 500;
+      color: #1f2937;
+    }
+
+    .upcomingStep {
+      color: #9ca3af;
+    }
+  }
+}
+
+.cardFooter {
+  margin-top: 12px;
+
+  .buttonOutline {
+    background-color: #f1f5f9;
+    color: #1f2937;
+    border: none;
+    padding: 8px 16px;
+    border-radius: 8px;
+    font-weight: 500;
+    font-size: 14px;
+    cursor: pointer;
+
+    &:hover {
+      background-color: #e2e8f0;
+    }
+  }
+}
+
+// Responsive design for mobile devices
+@media (max-width: 768px) {
+  .card {
+    padding: 1rem;
+  }
+
+  .cardHeader {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 0.5rem;
+
+    .proposalTitle {
+      font-size: 18px;
+    }
+
+    .proposalAuthor {
+      font-size: 12px;
+    }
+
+    .badge {
+      align-self: flex-start;
+    }
+  }
+
+  .cardMeta {
+    flex-direction: column;
+    gap: 1rem;
+
+    .metaItem {
+      .metaLabel {
+        font-size: 13px;
+      }
+
+      .metaValue {
+        font-size: 14px;
+      }
+    }
+  }
+
+  .progressSection {
+    .progressLabel {
+      font-size: 11px;
+    }
+
+    .progressSteps {
+      font-size: 11px;
+      flex-direction: column;
+      gap: 2px;
+
+      .currentStep,
+      .upcomingStep {
+        text-align: left;
+      }
+    }
+  }
+
+  .cardFooter {
+    .buttonOutline {
+      width: 100%;
+      text-align: center;
+    }
+  }
+}
