+++ app/governance/proposals/page.module.scss
@@ -63,23 +63,10 @@
   }
 }
 
-// .proposal-card {
-//   border: 1px solid #e5e7eb;
-//   border-radius: 0.75rem;
-//   padding: 1rem;
-//   margin-bottom: 2rem;
-//   background: white;
-// }
-
 .proposal-author {
   font-size: 0.875rem;
   color: #6b7280;
 }
-
-// .proposal-title {
-//   font-size: 1.25rem;
-//   font-weight: 600;
-// }
 
 .proposal-meta {
   display: flex;
@@ -109,14 +96,6 @@
   border-radius: 2rem;
   padding: 2rem;
 }
-
-// .proposal-list {
-//   display: flex;
-//   flex-direction: column;
-//   border: 1px solid #e5e7eb;
-//   border-radius: 0.75rem;
-//   overflow: hidden;
-// }
 
 .proposal-item {
   background: white;
@@ -185,34 +164,6 @@
   font-weight: 500;
 }
 
-// .vote-bar {
-//   grid-column: 3;
-//   grid-row: 2;
-//   margin-top: 0.25rem;
-//   width: 200px;
-//   height: 6px;
-//   justify-self: end;
-//   display: flex;
-//   border-radius: 4px;
-//   overflow: hidden;
-//   background: #d1d5db;
-// }
-
-// .vote-for {
-//   background-color: #10b981;
-//   height: 100%;
-// }
-
-// .vote-abstain {
-//   background-color: #6b7280;
-//   height: 100%;
-// }
-
-// .vote-against {
-//   background-color: #ef4444;
-//   height: 100%;
-// }
-
 .rejected {
   color: #ef4444;
 }
@@ -220,16 +171,6 @@
 .executed {
   color: #10b981;
 }
-
-// .loader {
-//   text-align: center;
-//   padding: 1.25rem 1.5rem;
-//   font-size: 0.875rem;
-//   color: #6b7280;
-//   background: white;
-//   border-top: 1px solid #e5e7eb;
-//   border-radius: 0 0 0.75rem 0.75rem;
-// }
 
 /* Responsive Design */
 @media (max-width: 768px) {
@@ -262,18 +203,6 @@
     grid-template-columns: 1fr;
     grid-template-rows: auto;
   }
-
-  // .status,
-  // .result,
-  // .votes,
-  // .vote-bar {
-  //   justify-self: start;
-  //   text-align: left;
-  // }
-
-  // .vote-bar {
-  //   width: 100%;
-  // }
 
   .primaryButton {
     width: 100%;
@@ -623,28 +552,6 @@
     }
   }
 }
-// .empty-state {
-//   text-align: center;
-//   padding: 4rem 2rem;
-//   background: var(--card-background, #f8f9fa);
-//   border-radius: 8px;
-//   border: 1px solid var(--border-color, #e9ecef);
-//   margin: 2rem 0;
-
-//   h3 {
-//     color: var(--text-primary, #333);
-//     font-size: 1.5rem;
-//     margin-bottom: 0.5rem;
-//     font-weight: 600;
-//   }
-
-//   p {
-//     color: var(--text-secondary, #666);
-//     font-size: 1rem;
-//     margin: 0;
-//     line-height: 1.5;
-//   }
-// }
 
 // Proposal List SCSS
 .proposal-list {
